+++ tools/iconprep/resize.js
@@ -1,5 +1,5 @@
 /**
- * This tool resizes given png/jpg/jpeg images to wanted sizes.
+ * This tool utilizes the GraphicsMagick to resize given png/jpg/jpeg images to wanted sizes.
  * Note that the generated images will be grouped under each 'size folder' inside the src image (or your dist) folder.
  *
  * =====
@@ -15,20 +15,24 @@
  *
  * @author Tim.Liu
  * @created 2013.10.23
+ * @updated 2014.06.13
  */
 
 var program = require('commander'),
 _ = require('underscore'),
 fs = require('fs'),
+mkdirp = require('mkdirp'),
 path = require('path'),
 colors = require('colors'),
-im = require('imagemagick');
+gm = require('gm');
+_.string = require('underscore.string');
 
 program
-	.version('0.1.0')
+	.version('0.1.1')
 	.usage('[options] <icon folder>')
-	.option('-S --sizes <16,32,64,.., array of sizes>', 'default on 16,32 (16x16 and 32x32)')
-	.option('-D --dist <path>', 'default to be the same as the src icon folder')
+	.option('-S --sizes <16,32,64,.., array of sizes>', 'default to 16,32 (16x16 and 32x32)')
+	.option('-K --keep', 'keep the original ones as unchanged copies [TBI]') //TBI
+	.option('-D --dist <path>', 'default to be <icon folder>/resized/')
 	.parse(process.argv);
 
 //check icon folder
@@ -37,6 +41,7 @@ if(!iconFolder) {
 	console.log('empty icon folder'.red);
 	return;
 }
+console.log('src:', '[', path.resolve(iconFolder).yellow, ']');
 
 //check on sizes
 if(program.sizes) {
@@ -44,29 +49,32 @@ if(program.sizes) {
 }else {
 	program.sizes = [16, 32];
 }
+console.log('sizing:', program.sizes.join(',').yellow);
 
 //start resizing
 fs.readdir(iconFolder, function(err, files){
 	if(err) throw err;
-	//create the size specific folders
-	_.each(program.sizes, function(size){
-		var sizeSpecFolder = path.join(program.dist || iconFolder, String(size));
-		if(!fs.existsSync(sizeSpecFolder)) fs.mkdirSync(sizeSpecFolder);
-	});
+
+	//create the resized icon dist folder
+	var distFolder = program.dist || path.join(iconFolder, 'resized');
+	if(!fs.existsSync(distFolder)) mkdirp.sync(distFolder);
+	console.log('dist:', '[', path.resolve(distFolder).yellow, ']');
+
 	//scan and convert the files
 	_.each(files, function(name){
 		var ext = path.extname(name);
 		if(!_.contains(['.png', '.jpg', '.jpeg'], ext)) return console.log(name, 'skipped...'.yellow);
 		_.each(program.sizes, function(size){
-			im.resize({
-				srcPath: path.join(iconFolder, name),
-				dstPath: path.join(program.dist || iconFolder, String(size), path.basename(name, ext) + '-' + String(size) + ext),
-				width: size,
-			}, function(err){
-				if(err) return console.log('ERROR'.red, err.message);
-				console.log('converted'.green, name, '==>'.green, size, 'x', size);
-			});
+			var rname = _.string.slugify(path.basename(name, ext)) + '-' + String(size) + ext;
+			//resize only the width while maintaining aspect ratio
+			gm(path.join(iconFolder, name))
+				.resize(size)
+				.write(path.join(distFolder, rname), function (err) {
+					if(err) return console.log('ERROR'.red, err.message);
+					console.log('converted'.green, name, '==>'.green, size, 'x', size, rname);
+				});
 			
 		})
 	});
+
 });